+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added NVPTX backend
 ### Fixed
 - Improved error handling
+- Fixed `-nostdlib` and `-nodefaultlibs` flags
 
 ## [3.3.2] - 2022-03-10
 ### Changed